requirements.txt
chromadb==0.4.22
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.0
langchain-google-genai==0.0.6
sentence-transformers==2.3.1
Django==5.0.1
python-dotenv==1.0.0